// prismicio-types.d.ts
type Simplify<T> = { [KeyType in keyof T]: T[KeyType] };

type PickContentRelationshipFieldData<
  TRelationship extends
    | prismic.CustomTypeModelFetchCustomTypeLevel1
    | prismic.CustomTypeModelFetchCustomTypeLevel2
    | prismic.CustomTypeModelFetchGroupLevel1
    | prismic.CustomTypeModelFetchGroupLevel2,
  TData extends Record<
    string,
    | prismic.AnyRegularField
    | prismic.GroupField
    | prismic.NestedGroupField
    | prismic.SliceZone
  >,
  TLang extends string,
> =
  // Content relationship fields
  {
    [TSubRelationship in Extract<
      TRelationship['fields'][number],
      prismic.CustomTypeModelFetchContentRelationshipLevel1
    > as TSubRelationship['id']]: ContentRelationshipFieldWithData<
      TSubRelationship['customtypes'],
      TLang
    >;
  } & { // Group
    [TGroup in Extract<
      TRelationship['fields'][number],
      | prismic.CustomTypeModelFetchGroupLevel1
      | prismic.CustomTypeModelFetchGroupLevel2
    > as TGroup['id']]: TData[TGroup['id']] extends prismic.GroupField<
      infer TGroupData
    >
      ? prismic.GroupField<
          PickContentRelationshipFieldData<TGroup, TGroupData, TLang>
        >
      : never;
  } & { // Other fields
    [TFieldKey in Extract<
      TRelationship['fields'][number],
      string
    >]: TFieldKey extends keyof TData ? TData[TFieldKey] : never;
  };

type ContentRelationshipFieldWithData<
  TCustomType extends
    | readonly (prismic.CustomTypeModelFetchCustomTypeLevel1 | string)[]
    | readonly (prismic.CustomTypeModelFetchCustomTypeLevel2 | string)[],
  TLang extends string = string,
> = {
  [ID in Exclude<
    TCustomType[number],
    string
  >['id']]: prismic.ContentRelationshipField<
    ID,
    TLang,
    PickContentRelationshipFieldData<
      Extract<TCustomType[number], { id: ID }>,
      Extract<prismic.Content.AllDocumentTypes, { type: ID }>['data'],
      TLang
    >
  >;
}[Exclude<TCustomType[number], string>['id']];
// ...
export interface NavbarDocumentDataLinksItem {
  /**
   * Link field in *Navbar → Links*
   *
   * - **Field Type**: Content Relationship
   * - **Placeholder**: *None*
   * - **API ID Path**: navbar.links[].link
   * - **Documentation**: https://prismic.io/docs/fields/content-relationship
   */
  link: ContentRelationshipFieldWithData<
    [
      {
        id: 'navlink';
        fields: [
          'main_link',
          'dropdown_label',
          { id: 'sublinks'; fields: ['link', 'description', 'icon'] },
        ];
      },
    ]
  >;
}

/**
 * Content for Navbar documents
 */
interface NavbarDocumentData {
  /**
   * Logo field in *Navbar*
   *
   * - **Field Type**: Image
   * - **Placeholder**: *None*
   * - **API ID Path**: navbar.logo
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/image
   */
  logo: prismic.ImageField<never>;

  /**
   * Logo Link field in *Navbar*
   *
   * - **Field Type**: Link
   * - **Placeholder**: *None*
   * - **API ID Path**: navbar.logo_link
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/link
   */
  logo_link: prismic.LinkField<
    string,
    string,
    unknown,
    prismic.FieldState,
    never
  >;

  /**
   * Links field in *Navbar*
   *
   * - **Field Type**: Group
   * - **Placeholder**: *None*
   * - **API ID Path**: navbar.links[]
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/repeatable-group
   */
  links: prismic.GroupField<Simplify<NavbarDocumentDataLinksItem>>;

  /**
   * CTA Link field in *Navbar*
   *
   * - **Field Type**: Link
   * - **Placeholder**: *None*
   * - **API ID Path**: navbar.cta_link
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/link
   */
  cta_link: prismic.Repeatable<
    prismic.LinkField<
      string,
      string,
      unknown,
      prismic.FieldState,
      'default' | 'secondary' | 'outline' | 'ghost'
    >
  >;

  /**
   * Mobile CTA Link field in *Navbar*
   *
   * - **Field Type**: Link
   * - **Placeholder**: *None*
   * - **API ID Path**: navbar.mobile_cta_link
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/link
   */
  mobile_cta_link: prismic.LinkField<
    string,
    string,
    unknown,
    prismic.FieldState,
    'default' | 'secondary' | 'outline' | 'ghost'
  > /**
   * Cart Title field in *Navbar*
   *
   * - **Field Type**: Text
   * - **Placeholder**: Cart
   * - **API ID Path**: navbar.cart_title
   * - **Tab**: Cart
   * - **Documentation**: https://prismic.io/docs/fields/text
   */;
  cart_title: prismic.KeyTextField;

  /**
   * Checkout Button Label field in *Navbar*
   *
   * - **Field Type**: Text
   * - **Placeholder**: Checkout
   * - **API ID Path**: navbar.checkout_button_label
   * - **Tab**: Cart
   * - **Documentation**: https://prismic.io/docs/fields/text
   */
  checkout_button_label: prismic.KeyTextField;

  /**
   * Clear Button Label field in *Navbar*
   *
   * - **Field Type**: Text
   * - **Placeholder**: Clear Cart
   * - **API ID Path**: navbar.clear_button_label
   * - **Tab**: Cart
   * - **Documentation**: https://prismic.io/docs/fields/text
   */
  clear_button_label: prismic.KeyTextField;

  /**
   * Total Label field in *Navbar*
   *
   * - **Field Type**: Text
   * - **Placeholder**: Total
   * - **API ID Path**: navbar.total_label
   * - **Tab**: Cart
   * - **Documentation**: https://prismic.io/docs/fields/text
   */
  total_label: prismic.KeyTextField;

  /**
   * Cart Disclaimer field in *Navbar*
   *
   * - **Field Type**: Rich Text
   * - **Placeholder**: Such as delivery times
   * - **API ID Path**: navbar.cart_disclaimer
   * - **Tab**: Cart
   * - **Documentation**: https://prismic.io/docs/fields/rich-text
   */
  cart_disclaimer: prismic.RichTextField;
}

/**
 * Navbar document from Prismic
 *
 * - **API ID**: `navbar`
 * - **Repeatable**: `false`
 * - **Documentation**: https://prismic.io/docs/content-modeling
 *
 * @typeParam Lang - Language API ID of the document.
 */
export type NavbarDocument<Lang extends string = string> =
  prismic.PrismicDocumentWithoutUID<
    Simplify<NavbarDocumentData>,
    'navbar',
    Lang
  >;

/**
 * Item in *Navlink → Sublinks*
 */
export interface NavlinkDocumentDataSublinksItem {
  /**
   * Link field in *Navlink → Sublinks*
   *
   * - **Field Type**: Link
   * - **Placeholder**: *None*
   * - **API ID Path**: navlink.sublinks[].link
   * - **Documentation**: https://prismic.io/docs/fields/link
   */
  link: prismic.LinkField<string, string, unknown, prismic.FieldState, never>;

  /**
   * Description field in *Navlink → Sublinks*
   *
   * - **Field Type**: Rich Text
   * - **Placeholder**: *None*
   * - **API ID Path**: navlink.sublinks[].description
   * - **Documentation**: https://prismic.io/docs/fields/rich-text
   */
  description: prismic.RichTextField;

  /**
   * Icon field in *Navlink → Sublinks*
   *
   * - **Field Type**: Image
   * - **Placeholder**: *None*
   * - **API ID Path**: navlink.sublinks[].icon
   * - **Documentation**: https://prismic.io/docs/fields/image
   */
  icon: prismic.ImageField<never>;
}
// ...
export type AllDocumentTypes =
  | CookieBannerDocument
  | FooterDocument
  | FourOhFourDocument
  | NavbarDocument
  | NavlinkDocument
  | PageDocument
  | TagDocument;